+++ components/display/message/message-item.tsx
@@ -125,21 +125,7 @@ export const MessageItem = ({
       console.log(error);
     }
   }
-  const onSubmit2 = async (values: z.infer<typeof formSchema>) => {
-    try {
-      const url = qs.stringifyUrl({
-        url: `${socketUrl}/${id}`,
-        query: socketQuery,
-      });
 
-      await axios.patch(url, values);
-
-      form.reset();
-      setIsEditing(false);
-    } catch (error) {
-      console.log(error);
-    }
-  }
   useEffect(() => {
     form.reset({
       content: content,