+++ type.ts
@@ -1,8 +1,5 @@
 
 import { Member, Message, Server, UserProfile ,AttachmentChannel, DirectMessage, DirectChannel, AttachmentDirect, Channel } from "@prisma/client";
-import { Server as NetServer, Socket } from "net";
-import { NextApiResponse } from "next";
-import { Server as SocketIOServer } from "socket.io";
 
 export type ServerWithMembersWithProfiles = 
                                     Server & {
@@ -19,14 +16,7 @@ export type DirectChannelWithProfile = DirectChannel & {
                                       memberOne: UserProfile,
                                       memberTwo: UserProfile
                                     }
-export type NextApiResponseServerIo = 
-    NextApiResponse & {
-        socket: Socket & {
-            server: NetServer & {
-            io: SocketIOServer;
-            };
-        };
-    };
+
 
 export type MessageWithMemberWithProfileWithFile = Message & {
         member: Member & {
@@ -35,18 +25,11 @@ export type MessageWithMemberWithProfileWithFile = Message & {
         AttachmentChannel: AttachmentChannel[];
       } 
 
-export type MessageWithProfileWithFile = DirectMessage & {
+export type DirectMessageWithProfileWithFile = DirectMessage & {
   userProfile: UserProfile ; 
   AttachmentDirect : AttachmentDirect[];
 };
          
-
-     
-export type MessageWithMemberWithProfileEU = Message & {
-    member: Member  
-    userProfile: UserProfile
-  };
-
  
 export type ServerWithChannels = Server & {
   Channel: Channel[]